styleguide and application in separate folders
