+++ package.json
@@ -10,6 +10,8 @@
     "styled-components": "^1.4.3"
   },
   "dependencies": {
+    "base-64": "^0.1.0",
+    "marked": "^0.3.6",
     "react": "^15.4.2",
     "react-dom": "^15.4.2"
   },
