+++ src/application/Application.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default () => (
+  <div>
+    <h2>Application goes here</h2>
+  </div>
+)
