+++ src/styleguide/components/contentContainer.js
